middleware/middleware/requirements.txt
backcall==0.2.0
blis==0.7.4
catalogue==2.0.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
cymem==2.0.5
decorator==5.0.9
filelock==3.0.12
Flask==2.0.1
Flask-Cors==3.0.10
fr-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_lg-3.1.0/fr_core_news_lg-3.1.0-py3-none-any.whl
# fr-dep-news-trf @ https://github.com/explosion/spacy-models/releases/download/fr_dep_news_trf-3.0.0/fr_dep_news_trf-3.1.0-py3-none-any.whl
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jaro-winkler==2.0.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
levenshtein==0.12.0
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
murmurhash==1.0.5
numpy==1.20.3
packaging==20.9
pandas==1.2.5
parso==0.8.2
pathy==0.5.2
pickleshare==0.7.5
Pillow==8.2.0
preshed==3.0.5
prompt-toolkit==3.0.19
protobuf==3.17.3
pydantic==1.7.4
Pygments==2.9.0
pyparsing==2.4.7
pysbd==0.3.4
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.1.0
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
scipy==1.7.0
seaborn==0.11.1
sentencepiece==0.1.91
six==1.16.0
smart-open==3.0.0
spacy==3.1.1
spacy-alignments==0.8.3
spacy-legacy==3.0.8
spacy-lookups-data==1.0.0
spacy-transformers==1.0.3
srsly==2.4.1
thinc==8.0.8
tokenizers==0.10.3
torch==1.8.1
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
transformers==4.5.1
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.5
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==2.0.1
redis==3.5.3
rq==1.9.0